feat: Extract business logic into custom hooks  - Create useProfileMenu hook for menu state management  - Create useLogoutOverlay hook for logout functionality  - Create useUserUtils hook for user data processing  - Simplify main component by removing inline logic  - Improve separation of concerns and reusability
