+++ hooks/use-profile-menu.ts
@@ -1,0 +1,62 @@
+import { useState, useRef, useEffect, useCallback } from "react";
+
+export function useProfileMenu() {
+  const [isProfileMenuOpen, setIsProfileMenuOpen] = useState(false);
+  const menuRef = useRef<HTMLDivElement>(null);
+  const buttonRef = useRef<HTMLButtonElement>(null);
+
+  // Close menu when clicking outside
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        menuRef.current &&
+        !menuRef.current.contains(event.target as Node) &&
+        buttonRef.current &&
+        !buttonRef.current.contains(event.target as Node)
+      ) {
+        setIsProfileMenuOpen(false);
+      }
+    };
+
+    if (isProfileMenuOpen) {
+      document.addEventListener('mousedown', handleClickOutside);
+    }
+
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [isProfileMenuOpen]);
+
+  // Close menu with Escape key
+  useEffect(() => {
+    const handleEscape = (event: KeyboardEvent) => {
+      if (event.key === 'Escape') {
+        setIsProfileMenuOpen(false);
+      }
+    };
+
+    if (isProfileMenuOpen) {
+      document.addEventListener('keydown', handleEscape);
+    }
+
+    return () => {
+      document.removeEventListener('keydown', handleEscape);
+    };
+  }, [isProfileMenuOpen]);
+
+  const toggleMenu = useCallback(() => {
+    setIsProfileMenuOpen(!isProfileMenuOpen);
+  }, [isProfileMenuOpen]);
+
+  const closeMenu = useCallback(() => {
+    setIsProfileMenuOpen(false);
+  }, []);
+
+  return {
+    isProfileMenuOpen,
+    menuRef,
+    buttonRef,
+    toggleMenu,
+    closeMenu,
+  };
+}
